+++ Board/src/main/java/com/study/domain/comment/CommentService.java
@@ -6,7 +6,7 @@ public interface CommentService {
 
 	public boolean registerComment(CommentDTO params);
 
-	public boolean deleteComment(Long id);
+	public boolean deleteComment(int id);
 
 	public List<CommentDTO> getCommentList(CommentDTO params);
